refactor: cleanup, move tray to its own mod
